Make Launch Configuration Editor Modal and clean needless code
Make the window Modal so as to avoid multi windows.
clean needless code about LaunchConfigurationEditor
