+++ SpatialGDK/Source/SpatialGDKEditor/Private/Utils/LaunchConfigEditorLayoutDetails.cpp
@@ -3,23 +3,12 @@
 #include "Utils/LaunchConfigEditorLayoutDetails.h"
 
 #include "SpatialGDKSettings.h"
-#include "Utils/LaunchConfigEditor.h"
+#include "Utils/LaunchConfigurationEditor.h"
 #include "DetailLayoutBuilder.h"
 
 TSharedRef<IDetailCustomization> FLaunchConfigEditorLayoutDetails::MakeInstance()
 {
 	return MakeShareable(new FLaunchConfigEditorLayoutDetails);
-}
-
-void FLaunchConfigEditorLayoutDetails::ForceRefreshLayout()
-{
-	if (MyLayout != nullptr)
-	{
-		TArray<TWeakObjectPtr<UObject>> Objects;
-		MyLayout->GetObjectsBeingCustomized(Objects);
-		ULaunchConfigurationEditor* Editor = Objects.Num() > 0 ? Cast<ULaunchConfigurationEditor>(Objects[0].Get()) : nullptr;
-		MyLayout->ForceRefreshDetails();
-	}
 }
 
 void FLaunchConfigEditorLayoutDetails::CustomizeDetails(IDetailLayoutBuilder& DetailBuilder)
