+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorLayoutDetails.h
@@ -14,9 +14,6 @@ public:
 	virtual void CustomizeDetails(IDetailLayoutBuilder& DetailBuilder) override;
 
 private:
-	void ForceRefreshLayout();
-
-private:
 	IDetailLayoutBuilder* CurrentLayout = nullptr;
 	TSharedPtr<IErrorReportingWidget> ProjectNameInputErrorReporting;
